requirements.txt
asgiref==3.6.0
Django==4.1.6
django-environ==0.9.0
environ==1.0
psycopg2==2.9.5
sqlparse==0.4.3

django-filter~=22.1